Proje/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
deep-translator==1.11.4
gtts
ffmpeg-python
pydub
srt
requests
python-dotenv
openai
pillow
numpy
scipy
librosa
whisper
moviepy
google-generativeai
deepL
torch
torchaudio
elevenlabs